chatbot-project/requirements.txt
FastAPI
uvicorn
pydantic
python-multipart
pytesseract
Pillow
numpy
scikit-learn
beautifulsoup4
nltk
requests
langchain
Groq